analysis/keyword_spellcheck/requirements.txt
pyspellchecker
langdetect
